Change uintptr_t to uint64_t where relevant

+++ obfuscator/lib/include/context.hpp
@@ -5,7 +5,6 @@
 #include <zasm/zasm.hpp>
 #include "instruction_data.hpp"
 
-
 namespace ObfuscatorLib
 {
 
@@ -16,10 +15,10 @@ class Context
   public:
     Context(Program& program);
 
-    void addInstructionData(Node* node, uintptr_t address, const InstructionDetail& detail);
+    void addInstructionData(Node* node, uint64_t address, const InstructionDetail& detail);
 
   private:
-    Program& program_;
+    Program&                    program_;
     std::deque<InstructionData> instructionDataPool_;
 };
