+++ obfuscator/lib/include/pe_loader.hpp
@@ -13,7 +13,7 @@ class PELoader
     ~PELoader();
 
     bool loadFile(const std::string& filePath);
-    bool extractFunction(const std::string& functionName, uintptr_t& address, std::vector<uint8_t>& functionData);
+    bool extractFunction(const std::string& functionName, uint64_t& address, std::vector<uint8_t>& functionData);
 
   private:
     std::vector<uint8_t> peData_;
